Move pages to module
